fix_bug: Entity delete with delete relations, Device
