+++ application/src/test/java/org/thingsboard/server/controller/BaseAlarmControllerTest.java
@@ -73,7 +73,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
 
         Mockito.reset(tbClusterService, auditLogService);
 
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         testNotifyEntityAllOneTime(alarm, alarm.getId(), alarm.getOriginator(),
                 tenantId, customerId, customerUserId, CUSTOMER_USER_EMAIL, ActionType.ADDED);
@@ -85,7 +85,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
 
         Mockito.reset(tbClusterService, auditLogService);
 
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         testNotifyEntityAllOneTime(alarm, alarm.getId(), alarm.getOriginator(),
                 tenantId, customerId, tenantAdminUserId, TENANT_ADMIN_EMAIL, ActionType.ADDED);
@@ -94,7 +94,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testUpdateAlarmViaCustomer() throws Exception {
         loginCustomerUser();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         Mockito.reset(tbClusterService, auditLogService);
 
@@ -110,7 +110,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testUpdateAlarmViaTenant() throws Exception {
         loginTenantAdmin();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         Mockito.reset(tbClusterService, auditLogService);
 
@@ -126,7 +126,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testUpdateAlarmViaDifferentTenant() throws Exception {
         loginTenantAdmin();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         alarm.setSeverity(AlarmSeverity.MAJOR);
         loginDifferentTenant();
@@ -143,7 +143,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testUpdateAlarmViaDifferentCustomer() throws Exception {
         loginCustomerUser();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         loginDifferentCustomer();
         alarm.setSeverity(AlarmSeverity.MAJOR);
@@ -160,7 +160,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testDeleteAlarmViaCustomer() throws Exception {
         loginCustomerUser();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         Mockito.reset(tbClusterService, auditLogService);
 
@@ -173,7 +173,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testDeleteAlarmViaTenant() throws Exception {
         loginTenantAdmin();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         Mockito.reset(tbClusterService, auditLogService);
 
@@ -186,7 +186,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testDeleteAlarmViaDifferentTenant() throws Exception {
         loginTenantAdmin();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         loginDifferentTenant();
 
@@ -202,7 +202,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testDeleteAlarmViaDifferentCustomer() throws Exception {
         loginCustomerUser();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         loginDifferentCustomer();
 
@@ -218,7 +218,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testClearAlarmViaCustomer() throws Exception {
         loginCustomerUser();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         Mockito.reset(tbClusterService, auditLogService);
 
@@ -235,7 +235,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testClearAlarmViaTenant() throws Exception {
         loginTenantAdmin();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         Mockito.reset(tbClusterService, auditLogService);
 
@@ -251,7 +251,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testAcknowledgeAlarmViaCustomer() throws Exception {
         loginCustomerUser();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         Mockito.reset(tbClusterService, auditLogService);
 
@@ -268,7 +268,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testClearAlarmViaDifferentCustomer() throws Exception {
         loginCustomerUser();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         loginDifferentCustomer();
 
@@ -284,7 +284,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testClearAlarmViaDifferentTenant() throws Exception {
         loginTenantAdmin();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         loginDifferentTenant();
 
@@ -300,7 +300,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testAcknowledgeAlarmViaDifferentCustomer() throws Exception {
         loginCustomerUser();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         loginDifferentCustomer();
 
@@ -316,7 +316,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
     @Test
     public void testAcknowledgeAlarmViaDifferentTenant() throws Exception {
         loginTenantAdmin();
-        Alarm alarm = createAlarm(TEST_ALARM_TYPE);
+        Alarm alarm = createAlarmForTest(TEST_ALARM_TYPE);
 
         loginDifferentTenant();
 
@@ -336,7 +336,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
         final int size = 10;
         for (int i = 0; i < size; i++) {
             createdAlarms.add(
-                    createAlarm(TEST_ALARM_TYPE + i)
+                    createAlarmForTest(TEST_ALARM_TYPE + i)
             );
         }
 
@@ -366,7 +366,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
 
         final int size = 10;
         for (int i = 0; i < size; i++) {
-            createAlarm(TEST_ALARM_TYPE + i);
+            createAlarmForTest(TEST_ALARM_TYPE + i);
         }
 
         loginDifferentCustomer();
@@ -389,12 +389,7 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
 
         String publicId = device.getCustomerId().toString();
 
-        Alarm alarm = Alarm.builder()
-                .originator(device.getId())
-                .status(AlarmStatus.ACTIVE_UNACK)
-                .severity(AlarmSeverity.CRITICAL)
-                .type("Test")
-                .build();
+        Alarm alarm = createAlarm("Test", device.getId());
 
         Mockito.reset(tbClusterService, auditLogService);
 
@@ -421,16 +416,8 @@ public abstract class BaseAlarmControllerTest extends AbstractControllerTest {
         Assert.assertTrue("Created alarm doesn't match the found one!", equals);
     }
 
-    private Alarm createAlarm(String type) throws Exception {
-        Alarm alarm = Alarm.builder()
-                .tenantId(tenantId)
-                .customerId(customerId)
-                .originator(customerDevice.getId())
-                .status(AlarmStatus.ACTIVE_UNACK)
-                .severity(AlarmSeverity.CRITICAL)
-                .type(type)
-                .build();
-
+    private Alarm createAlarmForTest(String type) throws Exception {
+        Alarm alarm = createAlarmWithTenant(tenantId, customerId, type, customerDevice.getId());
         alarm = doPost("/api/alarm", alarm, Alarm.class);
         Assert.assertNotNull(alarm);
